+++ src/util.c
@@ -6,15 +6,15 @@
 /*   By: fschmid <fschmid@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/13 16:49:55 by fschmid           #+#    #+#             */
-/*   Updated: 2023/01/13 16:55:36 by fschmid          ###   ########.fr       */
+/*   Updated: 2023/01/13 17:01:53 by fschmid          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
 
-int	getnext_whitespace(char *str)
+size_t	getnext_whitespace(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (str[++i] != '\0')
